修正：增加checksum校验

+++ nps/inc/hdr.h
@@ -38,7 +38,7 @@ typedef struct __attribute__((__packed__))
     uint8_t ihl:4;     // 因为字节序原因放在前面
     uint8_t version:4; // 因为字节序原因放在后面
     uint8_t tos;
-    uint8_t len;
+    uint16_t len;
     uint16_t identification;
     union
     {
